Complete debugging with the function of 'put' and 'get'. 1. The file seperator should be well considered when the operation is 'put' and 'get', in order to make the program run on Windows machine correct. Things to do: 1. Still need to redesign the subnet in sub-topology and top-topology. 2. Add a counter to execute the IC code several times in order to connect experimental data. 3. In intepreter, we need to design a counter to know what is the current counter in the loop.

+++ src/main/java/provisioning/engine/VEngine/EC2/EC2VEngine_ubuntu.java
@@ -63,6 +63,8 @@ public class EC2VEngine_ubuntu extends EC2VEngine implements VEngineCoreMethod, 
 		logger.debug("confFilePath: "+confFilePath);
 		try{
 		FileWriter fw = new FileWriter(confFilePath, false);
+		
+		boolean needConf = false;
 		////Configure for all top self connections
 		if(this.curVM.selfEthAddresses != null){
 			int count = 0;
@@ -83,6 +85,7 @@ public class EC2VEngine_ubuntu extends EC2VEngine implements VEngineCoreMethod, 
 					fw.write("route add -host "+localPrivateAddress+" dev "+linkName+"\n");
 					fw.flush();
 					entry.setValue(true);
+					needConf = true;
 				}
 			}
 		}
@@ -120,6 +123,7 @@ public class EC2VEngine_ubuntu extends EC2VEngine implements VEngineCoreMethod, 
 				fw.write("route del -net "+subnet+" netmask "+netmask+" dev "+linkName+"\n");
 				fw.write("route add -host "+remotePrivateAddress+" dev "+linkName+"\n");
 				fw.flush();
+				needConf = true;
 			}
 		}
 		
@@ -149,7 +153,6 @@ public class EC2VEngine_ubuntu extends EC2VEngine implements VEngineCoreMethod, 
 						}
 						curIndex++;
 					}
-					logger.debug("Get topconnection name "+linkName);
 					remotePubAddress = curTCP.peerTCP.belongingVM.publicAddress;
 					if(remotePubAddress == null){
 						curTCP.ethName = null;
@@ -164,6 +167,7 @@ public class EC2VEngine_ubuntu extends EC2VEngine implements VEngineCoreMethod, 
 				
 				///record the ethName
 				curTCP.ethName = linkName;
+				logger.debug("Get topconnection name "+linkName);
 				
 				fw.write("lp=`ifconfig eth0|grep 'inet addr'|awk -F'[ :]' '{print $13}'`\n");
 				fw.write("ip tunnel add "+linkName+" mode ipip remote "+remotePubAddress+" local $lp\n");
@@ -171,27 +175,29 @@ public class EC2VEngine_ubuntu extends EC2VEngine implements VEngineCoreMethod, 
 				fw.write("route del -net "+subnet+" netmask "+netmask+" dev "+linkName+"\n");
 				fw.write("route add -host "+remotePrivateAddress+" dev "+linkName+"\n");
 				fw.flush();
+				needConf = true;
 			}
 		}
 		fw.close();
 		
-
-		Thread.sleep(2000);
-		Shell shell = new SSH(curVM.publicAddress, 22, curVM.defaultSSHAccount, this.privateKeyString);
-		File file = new File(confFilePath);
-		new Shell.Safe(shell).exec(
-		  "cat > connection.sh && sudo bash connection.sh ",
-		  new FileInputStream(file),
-		  new NullOutputStream(), new NullOutputStream()
-		);
-		FileUtils.deleteQuietly(file);
-		new Shell.Safe(shell).exec(
-				  "rm connection.sh",
-				  null,
-				  new NullOutputStream(), new NullOutputStream()
-		);
+		if(needConf){
+			Shell shell = new SSH(curVM.publicAddress, 22, curVM.defaultSSHAccount, this.privateKeyString);
+			File file = new File(confFilePath);
+			new Shell.Safe(shell).exec(
+			  "cat > connection.sh && sudo bash connection.sh ",
+			  new FileInputStream(file),
+			  new NullOutputStream(), new NullOutputStream()
+			);
+			FileUtils.deleteQuietly(file);
+			new Shell.Safe(shell).exec(
+					  "rm connection.sh",
+					  null,
+					  new NullOutputStream(), new NullOutputStream()
+			);
+		}
 		
-		}catch (IOException | InterruptedException e) {
+		
+		}catch (IOException  e) {
 			e.printStackTrace();
 			logger.error(curVM.name +": "+ e.getMessage());
 			if(e.getMessage().contains("timeout: socket is not established")){   ////In this case, we give another chance to test.
