+++ src/main/java/provisioning/engine/VEngine/EGI/EGIVEngine_ubuntu.java
@@ -38,6 +38,7 @@ public class EGIVEngine_ubuntu extends EGIVEngine implements VEngineCoreMethod, 
         try {
             FileWriter fw = new FileWriter(confFilePath, false);
             
+            boolean needConf = false;
 	        ////Configure for all top self connections
 	    		if(this.curVM.selfEthAddresses != null){
 	    			int count = 0;
@@ -58,6 +59,7 @@ public class EGIVEngine_ubuntu extends EGIVEngine implements VEngineCoreMethod, 
 	    					fw.write("route add -host "+localPrivateAddress+" dev "+linkName+"\n");
 	    					fw.flush();
 	    					entry.setValue(true);
+	    					needConf = true;
 	    				}
 	    			}
 	    		}				
@@ -96,6 +98,7 @@ public class EGIVEngine_ubuntu extends EGIVEngine implements VEngineCoreMethod, 
                     fw.write("route del -net " + subnet + " netmask " + netmask + " dev " + linkName + "\n");
                     fw.write("route add -host " + remotePrivateAddress + " dev " + linkName + "\n");
                     fw.flush();
+                    needConf = true;
                 }
             }
 
@@ -126,7 +129,6 @@ public class EGIVEngine_ubuntu extends EGIVEngine implements VEngineCoreMethod, 
                             }
                             curIndex++;
                         }
-                        logger.debug("Get topconnection name " + linkName);
                         remotePubAddress = curTCP.peerTCP.belongingVM.publicAddress;
                         if (remotePubAddress == null) {
                             curTCP.ethName = null;
@@ -142,35 +144,36 @@ public class EGIVEngine_ubuntu extends EGIVEngine implements VEngineCoreMethod, 
 
                     ///record the ethName
                     curTCP.ethName = linkName;
+                    logger.debug("Get topconnection name " + linkName);
 
                     fw.write("lp=`ifconfig ens3|grep 'inet addr'|awk -F'[ :]' '{print $13}'`\n");
                     fw.write("ip tunnel add " + linkName + " mode ipip remote " + remotePubAddress + " local $lp\n");
                     fw.write("ifconfig " + linkName + " " + localPrivateAddress + " netmask " + netmask + "\n");
                     fw.write("route del -net " + subnet + " netmask " + netmask + " dev " + linkName + "\n");
                     fw.write("route add -host " + remotePrivateAddress + " dev " + linkName + "\n");
-                    
-
                     fw.flush();
+                    needConf = true;
                 }
             }
             fw.close();
 
-            Thread.sleep(2000);
-            Shell shell = new SSH(curVM.publicAddress, 22, curVM.defaultSSHAccount, this.privateKeyString);
-            File file = new File(confFilePath);
-            new Shell.Safe(shell).exec(
-                    "cat > connection.sh && sudo bash connection.sh ",
-                    new FileInputStream(file),
-                    new NullOutputStream(), new NullOutputStream()
-            );
-            FileUtils.deleteQuietly(file);
-//            new Shell.Safe(shell).exec(
-//                    "rm connection.sh",
-//                    null,
-//                    new NullOutputStream(), new NullOutputStream()
-//            );
+            if(needConf){
+	            Shell shell = new SSH(curVM.publicAddress, 22, curVM.defaultSSHAccount, this.privateKeyString);
+	            File file = new File(confFilePath);
+	            new Shell.Safe(shell).exec(
+	                    "cat > connection.sh && sudo bash connection.sh ",
+	                    new FileInputStream(file),
+	                    new NullOutputStream(), new NullOutputStream()
+	            );
+	            FileUtils.deleteQuietly(file);
+	            new Shell.Safe(shell).exec(
+	                    "rm connection.sh",
+	                    null,
+	                    new NullOutputStream(), new NullOutputStream()
+	            );
+            }
 
-        } catch (IOException | InterruptedException e) {
+        } catch (IOException  e) {
             e.printStackTrace();
             logger.error(curVM.name + ": " + e.getMessage());
             if (e.getMessage().contains("timeout: socket is not established")) {   ////In this case, we give another chance to test.
@@ -192,16 +195,6 @@ public class EGIVEngine_ubuntu extends EGIVEngine implements VEngineCoreMethod, 
                         throw new RuntimeException(e1);
                     }
                 }
-/*                try {
-                      Shell shell = new SSH(curVM.publicAddress, 22, curVM.defaultSSHAccount, this.privateKeyString);
-//                    new Shell.Safe(shell).exec(
-//                            "rm connection.sh",
-//                            null,
-//                            new NullOutputStream(), new NullOutputStream()
-//                    );
-                } catch (IOException e1) {
-                    throw new RuntimeException(e1);
-                }*/
             }
         }
     }
